rename scores class to leaderboard Fixes #2
